refactor(executor): extract Codex JSON-RPC transport
